+++ kernel/boot/gdt.c
@@ -30,7 +30,7 @@ void gdt_set_gate(struct gdt_entry *gdt, int num, uint64_t base, uint32_t limit,
     gdt[num].base_high = (base >> 24) & 0xFF;
 }
 
-void gdt_load(struct gdt_entry *gdt, size_t n_entries) {
+void gdt_load(struct gdt_entry *gdt, uint64_t n_entries) {
     struct gdt_ptr gp = {
         .limit = (sizeof(struct gdt_entry) * n_entries) - 1,
         .base = (uint64_t) gdt,
